+++ src/hooks/useUserData.tsx
@@ -1,0 +1,32 @@
+import { useState, useEffect } from 'react'
+
+import AXIOS from '@/lib/axios'
+
+type UserData = {
+  email: string
+  nickname: string
+  profileImageUrl: string
+}
+
+export function useUserData() {
+  const [userData, setUserData] = useState<UserData | null>(null)
+
+  useEffect(() => {
+    const accessToken = localStorage.getItem('accessToken')
+    if (accessToken) {
+      AXIOS.get('/users/me', {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      })
+        .then((res) => {
+          setUserData(res.data)
+        })
+        .catch((err) => {
+          console.error('Error fetching user data:', err.response.data.message)
+        })
+    }
+  }, [])
+
+  return userData
+}
